Use uint32_t type for RewardsNotificationID, to avoid JavaScript overflow

+++ components/brave_rewards/browser/rewards_notifications_service.h
@@ -20,7 +20,7 @@ class RewardsNotificationsService : public KeyedService {
   RewardsNotificationsService();
   ~RewardsNotificationsService() override;
 
-  typedef uint64_t RewardsNotificationID;
+  typedef uint32_t RewardsNotificationID;
   typedef int64_t RewardsNotificationTimestamp;
   typedef std::vector<std::string> RewardsNotificationArgs;
